+++ src/logic/systems/system_list.h
@@ -11,7 +11,7 @@ public: \
     NAME(entt::registry& reg) : System(reg) {}; \
     void preinit() override; \
     void init() override; \
-    void tick(float dt) override; \
+    void tick(double dt) override; \
     void finish() override; \
     const char* name() override { return #NAME; }; \
 };
@@ -19,15 +19,25 @@ public: \
 namespace dy {
 
 DEFINE_SYSTEM(CameraSys)
+
 DEFINE_SYSTEM(TimeSys)
+
 DEFINE_SYSTEM(DevMenuSys)
+
 DEFINE_SYSTEM(ChunkGenerationSys)
+
 DEFINE_SYSTEM(PatherSys)
+
 DEFINE_SYSTEM(BasicNeedsSys)
+
 DEFINE_SYSTEM(HarvestSys)
+
 DEFINE_SYSTEM(EatSys)
+
 DEFINE_SYSTEM(ActionBarSys)
+
 DEFINE_SYSTEM(MapConfiguratorSys)
+
 DEFINE_SYSTEM(MapEditorSys)
 
 }
