+++ src/logic/systems/ui_render.h
@@ -38,7 +38,7 @@ public:
     }
 
     void init() override;
-    void tick(float dt) override;
+    void tick(double dt) override;
     
     void createOrResizeBuffer(vk::Buffer& buffer, vk::DeviceMemory& buffer_memory, vk::DeviceSize& p_buffer_size, size_t new_size,vk::BufferUsageFlagBits usage);
